Refactor preview prop resolution and sync controls
