+++ playground/src/views/core/index.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { computed, ref } from 'vue'
-import type { LabComponentDefinition } from '@/library/catalog'
+import type { LabComponentDefinition, PlaygroundPropValue } from '@/library/catalog'
 import { getComponentDefinition } from '@/library/catalog'
 import Controls from './controls/index.vue'
 import NotFound from './notFound.vue'
@@ -12,22 +12,18 @@ const props = defineProps<{
 
 const definition = computed<LabComponentDefinition | undefined>(() => getComponentDefinition(props.componentId))
 
-const resolvedComponentProps = ref<Record<string, unknown>>({})
-
-const handleResolvedPropsUpdate = (nextResolvedProps: Record<string, unknown>) => {
-  resolvedComponentProps.value = nextResolvedProps
-}
+const componentProps = ref<Record<string, PlaygroundPropValue>>({})
 </script>
 
 <template>
   <div v-if="definition" class="grid gap-8 lg:grid-cols-[minmax(0,1.2fr)_22rem]">
     <Preview
       :definition="definition"
-      :resolved-component-props="resolvedComponentProps"
+      :component-props="componentProps"
     />
     <Controls
       :definition="definition"
-      @update:resolved-props="handleResolvedPropsUpdate"
+      v-model:props="componentProps"
     />
   </div>
   <NotFound v-else />
