+++ playground/src/views/core/preview.vue
@@ -3,12 +3,12 @@ import { computed, defineAsyncComponent, watch } from 'vue'
 import type { AsyncComponentLoader } from 'vue'
 import { useI18n } from 'vue-i18n'
 import Spinner from '@library/components/Spinner.vue'
-import type { LabComponentDefinition, LocaleCopy } from '@/library/catalog'
+import type { LabComponentDefinition, LocaleCopy, PlaygroundPropValue } from '@/library/catalog'
 import type { SupportedLocale } from '@/i18n'
 
 const props = defineProps<{
   definition: LabComponentDefinition
-  resolvedComponentProps: Record<string, unknown>
+  componentProps: Record<string, PlaygroundPropValue>
 }>()
 
 const { t, locale } = useI18n()
@@ -18,12 +18,49 @@ const localize = (copy: LocaleCopy) => copy[(locale.value as SupportedLocale)] ?
 const localizedName = computed(() => localize(props.definition.name))
 const localizedDescription = computed(() => localize(props.definition.description))
 
+const resolveComponentProps = (flatProps: Record<string, PlaygroundPropValue>) => {
+  const expanded: Record<string, unknown> = {}
+
+  Object.entries(flatProps).forEach(([path, value]) => {
+    const segments = path.split('.')
+    if (segments.length === 0) {
+      return
+    }
+
+    let target: Record<string, unknown> = expanded
+    for (let index = 0; index < segments.length - 1; index += 1) {
+      const segment = segments[index]
+      if (!segment) {
+        return
+      }
+
+      const existing = target[segment]
+      if (!existing || typeof existing !== 'object' || Array.isArray(existing)) {
+        target[segment] = {}
+      }
+
+      target = target[segment] as Record<string, unknown>
+    }
+
+    const lastSegment = segments[segments.length - 1]
+    if (!lastSegment) {
+      return
+    }
+
+    target[lastSegment] = value as unknown
+  })
+
+  return expanded
+}
+
 const previewComponent = computed(() => {
   const loader = props.definition.preview ?? props.definition.component
   return defineAsyncComponent(loader as AsyncComponentLoader<any>)
 })
 
 const hasPreview = computed(() => previewComponent.value !== null)
+
+const resolvedComponentProps = computed(() => resolveComponentProps(props.componentProps))
 
 watch(
   [localizedName, () => String(t('app.title'))],
